+++ features/fixtures/plain/app/report_modification/initiators/unhandled_on_error.rb
@@ -1,0 +1,11 @@
+require 'bugsnag'
+require './app'
+
+configure_basics
+add_at_exit
+
+def run(callback)
+  Bugsnag.add_on_error(callback)
+
+  raise RuntimeError.new "Oh no"
+end