+++ spec/fixtures/crashes/file_with_long_lines.rb
@@ -1,0 +1,7 @@
+# rubocop:disable Layout/LineLength
+def a_super_long_function_name_that_would_be_really_impractical_to_use_but_luckily_this_is_just_for_a_test_to_prove_we_can_handle_really_long_lines_of_code_that_go_over_200_characters_and_some_more_padding
+  puts 'This is a shorter string'
+  puts 'A more realistic example of when a line would be really long is long strings such as this one, which extends over the 200 character limit by containing a lot of excess words for padding its length so that it is super long'
+  puts 'and another shorter string for comparison'
+end
+# rubocop:enable Layout/LineLength